refactor(execution): migrate 19 internal imports to public API paths
- models.verdict/verdict_types → models (VerdictRecord, VerdictValue)
- models.review_runner → models (AgentOptions)
- config.role_policy → config (get_role_output_contract, get_role_section)
- config.agent_preset → config (resolve_effective_agent_options)
- config.role_gates → config (7 gate configs)
- agents.backends.codeagent_config → agents (sync_models_json)
- environment.session_registry → environment (SessionRegistryService)
- clients.store_context → clients (get_store)
- exceptions.runtime_errors → exceptions (GitHubAPIError)

12 files, net -7 lines. mypy clean, 17/17 export tests pass.
