+++ src/vibe3/execution/capacity_service.py
@@ -14,7 +14,7 @@ from typing import TYPE_CHECKING
 from loguru import logger
 
 from vibe3.clients import BackendProtocol, SQLiteClient
-from vibe3.environment.session_registry import SessionRegistryService
+from vibe3.environment import SessionRegistryService
 from vibe3.models import OrchestraConfig
 
 if TYPE_CHECKING:
