+++ src/vibe3/execution/codeagent_runner.py
@@ -20,8 +20,7 @@ if TYPE_CHECKING:
         CodeagentResult,
     )
     from vibe3.exceptions.error_severity import ErrorSeverity
-from vibe3.config import VibeConfig
-from vibe3.config.role_policy import get_role_section
+from vibe3.config import VibeConfig, get_role_section
 from vibe3.execution.codeagent_support import resolve_command_agent_options
 from vibe3.execution.execution_lifecycle import (
     execution_prefix,
@@ -29,8 +28,7 @@ from vibe3.execution.execution_lifecycle import (
 )
 from vibe3.execution.noop_gate import apply_unified_noop_gate
 from vibe3.execution.session_service import load_session_id
-from vibe3.models import ExecutionRequest
-from vibe3.models.review_runner import AgentOptions
+from vibe3.models import AgentOptions, ExecutionRequest
 from vibe3.services import HandoffService, format_agent_actor
