+++ src/vibe3/execution/issue_role_support.py
@@ -8,8 +8,13 @@ from typing import Any, Callable, Iterable
 
 from vibe3.clients import SQLiteClient
 from vibe3.execution.role_interfaces import IssueRoleSyncSpec
-from vibe3.models import ExecutionRequest, IssueInfo, SessionRole, WorktreeRequirement
-from vibe3.models.review_runner import AgentOptions
+from vibe3.models import (
+    AgentOptions,
+    ExecutionRequest,
+    IssueInfo,
+    SessionRole,
+    WorktreeRequirement,
+)
 from vibe3.roles.definitions import IssueRoleSyncSpec as IssueRoleSyncSpecImpl
