+++ src/vibe3/execution/noop_gate.py
@@ -7,9 +7,8 @@ from loguru import logger
 
 from vibe3.agents import ExecutionRole
 from vibe3.clients import SQLiteClient
-from vibe3.config.role_policy import get_role_output_contract
-from vibe3.models.verdict import VerdictRecord
-from vibe3.models.verdict_types import VerdictValue
+from vibe3.config import get_role_output_contract
+from vibe3.models import VerdictRecord, VerdictValue
 from vibe3.services import get_role_block_function
 
 # Loop prevention constants
